+++ src/disco/topo/fd_topo.h
@@ -253,7 +253,7 @@ struct fd_topo_tile {
         uchar tip_distribution_program_addr[ 32 ];
         uchar tip_payment_program_addr[ 32 ];
         uchar tip_distribution_authority[ 32 ];
-        ulong commission_bps;
+        uint  commission_bps;
         char  identity_key_path[ PATH_MAX ];
         char  vote_account_path[ PATH_MAX ]; /* or pubkey is okay */
       } bundle;
